backend/djangoapi_backend/requirements.txt
django
djangorestframework
django-cors-headers
django-environ
gunicorn
celery
redis
drf-yasg